requirements.txt
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.15.0
datasets>=1.18.0
opencv-python>=4.5.5
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0
tqdm>=4.62.0
scikit-learn>=1.0.0
jiwer>=2.3.0
pandas>=1.3.5 